requirements.txt
PyQt6==6.9.0
requests==2.32.4
matplotlib==3.8.4
sounddevice==0.5.2
soundfile==0.13.1
numpy==1.26.4
pygame==2.6.1
pyserial==3.5
reportlab==4.4.2
scipy==1.13.1
python-dotenv==1.0.0